requirements.txt
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
python-dotenv==1.0.0
transformers==4.36.0
torch==2.1.0
huggingface_hub==0.24.1
pyttsx3==2.90
dateparser==1.1.8
parsedatetime==2.6
python-dateutil==2.8.2
pydub==0.25.1
sounddevice==0.4.6
numpy==1.24.3
scipy==1.10.1
pandas==2.0.3
openai-whisper
